requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
doc8==0.8.0
docutils==0.14
emoji==0.5.2
ephem==3.7.6.0
future==0.17.1
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pbr==5.2.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
PySocks==1.7.0
python-telegram==0.10.0
python-telegram-bot==11.1.0
restructuredtext-lint==1.3.0
six==1.12.0
stevedore==1.30.1
typed-ast==1.3.5
wrapt==1.11.1
